Change all memory and disk usage to uint64
This will get tests passing on Windows 32bit.
May be a bit overkill but tests are green.

[#57413668]

+++ src/cf/api/app_summary.go
@@ -47,12 +47,12 @@ type StatsApiResponse map[string]InstanceStatsApiResponse
 
 type InstanceStatsApiResponse struct {
 	Stats struct {
-		DiskQuota int `json:"disk_quota"`
-		MemQuota  int `json:"mem_quota"`
+		DiskQuota uint64 `json:"disk_quota"`
+		MemQuota  uint64 `json:"mem_quota"`
 		Usage     struct {
 			Cpu  float64
-			Disk int
-			Mem  int
+			Disk uint64
+			Mem  uint64
 		}
 	}
 }
